+++ src/app/program/edit/edit-event/edit-event.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { EventService } from '../../../services/event.service';
+import { EventService } from '../../../services/event/event.service';
 import { EventDto } from '../../../models/models';
 
 @Component({
