+++ src/app/program/view/program.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { EventCategoryDto, EventDto } from '../../models/models';
-import { EventService } from '../../services/event.service';
-import { DateConverterService } from '../../services/date-converter.service';
-import { EventCategoriesService } from '../../services/event-categories.service';
+import { EventService } from '../../services/event/event.service';
+import { DateConverterService } from '../../services/shared/date-converter.service';
+import { EventCategoriesService } from '../../services/event/event-categories.service';
 
 @Component({
   selector: 'app-footer',
