+++ src/app/services/event/event-info.service.spec.ts
