+++ src/app/services/event/event-info.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { EventInfoDto } from '../models/models';
+import { EventInfoDto } from '../../models/models';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
@@ -9,19 +9,19 @@ import { Observable } from 'rxjs';
 export class EventInfoService {
   constructor(private httpClient: HttpClient) {}
 
-  getEventInfoById(eventInfoId: number): Observable<EventInfoDto> {
-    return this.httpClient.get<EventInfoDto>('api/eventInfos/' + eventInfoId);
+  getEventInfoById(id: number): Observable<EventInfoDto> {
+    return this.httpClient.get<EventInfoDto>(`api/eventInfos/${id}`);
   }
 
   updateEventInfo(id: number, eventInfo: EventInfoDto) {
-    return this.httpClient.put<EventInfoDto>('api/eventInfos/' + id, eventInfo);
+    return this.httpClient.put<EventInfoDto>(`api/eventInfos/${id}`, eventInfo);
   }
 
   createEventInfo(eventInfo: EventInfoDto) {
     return this.httpClient.post<EventInfoDto>('api/eventInfos', eventInfo);
   }
 
-  deleteEventInfoById(eventInfoId: number) {
-    return this.httpClient.delete<Response>('api/eventInfos/' + eventInfoId);
+  deleteEventInfoById(id: number) {
+    return this.httpClient.delete<Response>(`api/eventInfos/${id}`);
   }
 }
