+++ src/app/services/event/event.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { EventDto } from '../models/models';
+import { EventDto } from '../../models/models';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
@@ -13,22 +13,19 @@ export class EventService {
     return this.httpClient.get<EventDto[]>('api/events');
   }
 
-  getEventByEventId(eventId: number): Observable<EventDto> {
-    return this.httpClient.get<EventDto>('api/events/' + eventId);
+  getEventByEventId(id: number): Observable<EventDto> {
+    return this.httpClient.get<EventDto>(`api/events/${id}`);
   }
 
   updateEvent(id: number, eventDto: EventDto) {
-    return this.httpClient.put<EventDto>('api/events/' + id, {
-      id,
-      eventDto,
-    });
+    return this.httpClient.put<EventDto>(`api/events/${id}`, { eventDto });
   }
 
   createEvent(eventDto: EventDto) {
     return this.httpClient.post<EventDto>('api/events', eventDto);
   }
 
-  deleteEventByEventId(eventId: number) {
-    return this.httpClient.delete<Response>('api/events/' + eventId);
+  deleteEventByEventId(id: number) {
+    return this.httpClient.delete<Response>(`api/events/${id}`);
   }
 }
