+++ src/app/voting/add-edit/voting-add-edit.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import { EventDto, VotingDto } from '../../models/models';
-import { EventService } from '../../services/event.service';
+import { EventService } from '../../services/event/event.service';
 import { VotingService } from '../../services/voting.service';
 
 @Component({
